% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{add_call}
\alias{add_call}
\title{Adds a Call}
\usage{
add_call(id, new_call, description = NULL, env = rlang::caller_env())
}
\arguments{
\item{id}{Identifier}

\item{new_call}{Call}

\item{description}{Description
The description is added as comment before the call.}

\item{env}{Environment where to store the call}
}
\description{
Adds a Call
}
