% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{add_call_template}
\alias{add_call_template}
\title{Adds a Call Template}
\usage{
add_call_template(
  id,
  call,
  description = NULL,
  defaults = NULL,
  env = rlang::caller_env()
)
}
\arguments{
\item{id}{Identifier}

\item{call}{Call}

\item{description}{Description
The description is added as comment before the call.}

\item{defaults}{List of default values}

\item{env}{Environment where to store the call template}
}
\description{
Adds a Call Template
}
