% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_file_exists.R
\name{assert_file_exists}
\alias{assert_file_exists}
\title{Asserts That a File Exists}
\usage{
assert_file_exists(filename)
}
\arguments{
\item{filename}{character, file path}
}
\description{
Throws an error if a file does not exist
}
