% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptr.R
\name{copy_script}
\alias{copy_script}
\title{Copy a script to the clipboard}
\usage{
copy_script(name)
}
\arguments{
\item{name}{A character string.}
}
\value{
A message indicating that the script has been copied to the clipboard.
}
\description{
Copy a script to the clipboard
}
