% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customscript.R
\name{customscript}
\alias{customscript}
\title{Create a Modelling Template in RStudio}
\usage{
customscript()
}
\description{
This function loads a new file fir data science analysis
using a pre-set outline template.
}
