% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictConfig.R
\name{dictConfig}
\alias{dictConfig}
\title{Takes the Logging Configuration From a List}
\usage{
dictConfig(config)
}
\arguments{
\item{config}{configurations as a list}
}
\description{
R adaptation of the Python logging.config.dictConfig function.
See \url{https://docs.python.org/3.6/library/logging.config.html#dictionary-schema-details}
for schema details.
}
\examples{
\dontrun{
# from an external json file
config <- rjson::fromJSON(file = "log_config.json")
dictConfig(config)

# use the internal log_config list
dictConfig(log_config)
}

}
