% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensure_requisite_folders.R
\name{ensure_requisite_folders}
\alias{ensure_requisite_folders}
\title{Creates Folder if it Doesn't Exist}
\usage{
ensure_requisite_folders(dir)
}
\arguments{
\item{dir}{filepath to required folder}
}
\description{
Creates Folder if it Doesn't Exist
}
