% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{execute}
\alias{execute}
\title{Generate a Script}
\usage{
execute(code, filename)
}
\arguments{
\item{code}{Code for creating the script}

\item{file}{File to write the script to}
}
\description{
Generate a Script
}
