% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_conflicts}
\alias{get_conflicts}
\title{Get Conflicts}
\usage{
get_conflicts(pkgs)
}
\arguments{
\item{pkgs}{a character vector of package names}
}
\value{
a data frame
}
\description{
Get Conflicts
}
