% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_library_search}
\alias{get_library_search}
\title{Get Script Search Path}
\usage{
get_library_search(x)
}
\arguments{
\item{x}{a character vector of package names}
}
\value{
a character vector of package names
}
\description{
Get Script Search Path
}
