% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{get_unused_objs}
\alias{get_unused_objs}
\title{Get Unused Objects}
\usage{
get_unused_objs(code)
}
\arguments{
\item{code}{a data frame as included in the $code of any ScriptR object}
}
\value{
a data frame
}
\description{
Get Unused Objects
}
