% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{insert_calls}
\alias{insert_calls}
\title{Inserts Calls to a Script}
\usage{
insert_calls(input, ids)
}
\arguments{
\item{input}{Input object

The specified object is passed via a pipe to the first call.}

\item{ids}{Calls

A list of call ids is expected.}
}
\description{
Inserts Calls to a Script
}
