% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{insert_chunks}
\alias{insert_chunks}
\title{Inserts Chunks into a Script}
\usage{
insert_chunks(ids)
}
\arguments{
\item{ids}{Chunk ids

A list of chunk ids is expected.}
}
\description{
Inserts Chunks into a Script
}
