% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_named_list.R
\name{is_named_list}
\alias{is_named_list}
\title{Checks Whether a List is a Uniquely Named List}
\usage{
is_named_list(l)
}
\arguments{
\item{l}{a list}
}
\value{
boolean
}
\description{
Checks Whether a List is a Uniquely Named List
}
