% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{log_config}
\alias{log_config}
\title{Default Log Config Dictionary}
\format{A list with the formatting for the logging config}
\usage{
log_config
}
\description{
This is the default list/dictionary for start_logging if a config file is not given. It should be used as a template for the json files for the logging config.
}
\keyword{datasets}
