% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_function_name.R
\name{log_function_name}
\alias{log_function_name}
\title{Wrapper: Logs the Function Name}
\usage{
log_function_name(func)
}
\arguments{
\item{func}{function to wrap}
}
\description{
Logs the function name at the INFO level before running the function.
}
