% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfile.R
\name{logfile}
\alias{logfile}
\alias{exists_logfile}
\alias{get_logfile}
\title{Logfile Information}
\usage{
exists_logfile()

get_logfile()
}
\description{
Obtain information regarding the logfile
}
\details{
exists_logfile() checks whether the logfile exists
get_logfile() obtains the name of the logfile
}
