% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmisc.R
\name{logmisc}
\alias{logmisc}
\title{Logs Misc Data Structures}
\usage{
logmisc(x, loglevel = "INFO")
}
\arguments{
\item{x}{any object to be logged}

\item{loglevel}{level to log the message}
}
\description{
Allows for logging of data structures such as dataframes, etc.
}
\examples{
logmisc(iris)
logmisc(summary(iris))

}
