% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{parse_script}
\alias{parse_script}
\title{Parse Scripts}
\usage{
parse_script(path)
}
\arguments{
\item{path}{a character vector of source files.}
}
\value{
a ScriptR object
}
\description{
Parse Scripts
}
