% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cmd_args.R
\name{print_cmd_args}
\alias{print_cmd_args}
\title{Prints Command Line Args}
\usage{
print_cmd_args(args)
}
\arguments{
\item{args}{list of command line args, generated from argsparse::parse_args}
}
\description{
Prints out all command line arguments
}
\examples{
scriptR::start_logging()

option_list <- list(make_option("--test_opt", help = "this is a parameter"))
opt <- parse_args(OptionParser(description = "This is a test", option_list = option_list))

print_cmd_args(opt)

}
