% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_args.R
\name{process_args}
\alias{process_args}
\title{Process Command Line Args and Initiate Logging}
\usage{
process_args(parser, description, ...)
}
\arguments{
\item{parser}{an argparse object}

\item{description}{description of script}

\item{...}{additional options passed to parse_args() call}
}
\value{
a list parsed args
}
\description{
Processes command line args, logs script description & arguments passed in. Returns list of arg values.
}
\examples{

description <- "TEST"
parser <- ArgumentParser(description = description)
parser$add_argument("--test_opt", help = "This is a test")
parser$add_argument("--log", default = NULL, help = "Name of log config file [$(default)]")

args <- process_options(parser = parser, description = description, ...)

}
