% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.in.R
\name{read.in}
\alias{read.in}
\title{Opens a File}
\usage{
read.in(filename, f = data.table::fread, ...)
}
\arguments{
\item{filename}{character, file path}

\item{f}{function, reads in the file}

\item{...}{additional arguments passed to f}
}
\description{
Checks whether a file exists and opens it using the function you request
}
\examples{
read.in("test.csv")
read.in("dat.csv", read.table, header = TRUE)

}
