% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptR-package.R
\docType{package}
\name{scriptR-package}
\alias{scriptR}
\alias{scriptR-package}
\title{scriptR: What the Package Does (Title Case)}
\description{
This package parses scripts into a format that is suitable for analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/acastroaraujo/scriptR}
  \item Report bugs at \url{https://github.com/acastroaraujo/scriptR/issues}
}

}
\keyword{internal}
