% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_level_global.R
\name{set_level_global}
\alias{set_level_global}
\title{Globally Set Logging Level}
\usage{
set_level_global(level)
}
\arguments{
\item{level}{logging level, see logging::loglevels for available options}
}
\description{
Set the logging level across all handlers
}
\examples{
set_level_global("CRITICAL")

}
