% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setLevel_global.R
\name{setLevel_global}
\alias{setLevel_global}
\title{Globally Set Logging Level}
\usage{
setLevel_global(level)
}
\arguments{
\item{level}{logging level, see logging::loglevels for available options}
}
\description{
Set the logging level across all handlers
}
\examples{
setLevel_global("CRITICAL")

}
