% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_logging.R
\name{start_logging}
\alias{start_logging}
\title{Initiates a Logging Session}
\usage{
start_logging(make_file = FALSE, file_name)
}
\arguments{
\item{make_file}{should a log file be created}

\item{file_name}{name of the log file, if none given will save to default (R_[timestamp])}
}
\description{
Logs messages to command line, with option to create a log file.
}
\examples{
start_logging(FALSE)
start_logging(TRUE, "test.log")

}
