% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calls.R
\name{add_call}
\alias{add_call}
\title{Adds a Call}
\usage{
add_call(id, call, env = rlang::caller_env())
}
\arguments{
\item{id}{Identifier}

\item{call}{Call}

\item{env}{Environment where to store the call}
}
\description{
Adds a Call
}
