% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{add_chunk}
\alias{add_chunk}
\title{Adds a Chunk}
\usage{
add_chunk(id, chunk, env = rlang::caller_env())
}
\arguments{
\item{id}{Identifier}

\item{chunk}{Chunk}

\item{env}{Environment where to store the chunk}
}
\description{
Adds a Chunk
}
