% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks.R
\name{add_chunk_template}
\alias{add_chunk_template}
\title{Adds a Chunk Template}
\usage{
add_chunk_template(
  id,
  chunk,
  description = NULL,
  defaults = NULL,
  env = rlang::caller_env()
)
}
\arguments{
\item{id}{Identifier}

\item{chunk}{Chunk}

\item{description}{Description
The description is added as comment before the chunk.}

\item{defaults}{List of default values}

\item{env}{Environment where to store the chunk template}
}
\description{
Adds a Chunk Template
}
