% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_cols_in.R
\name{assert_cols_in}
\alias{assert_cols_in}
\title{Asserts That Columns are in Data Frame}
\usage{
assert_cols_in(d, cols)
}
\arguments{
\item{d}{data.frame}

\item{cols}{character vector of columns to check for}
}
\description{
Asserts That Columns are in Data Frame
}
\examples{
assert_cols_in(iris, c("Petal.Width", "Petal.Length")) # cols in, will do nothing
\dontrun{
assert_cols_in(iris, c("x", "y")) # cols not in, will assert error
}

}
