% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictConfig.R
\name{dictConfig}
\alias{dictConfig}
\title{Takes the logging configuration from a list}
\usage{
dictConfig(config)
}
\arguments{
\item{config}{configurations as a list}
}
\description{
R adaptation of the Python logging.config.dictConfig function. See \url{https://docs.python.org/3.6/library/logging.config.html#dictionary-schema-details} for schema details.
}
\examples{
config <- rjson::fromJSON(file = "log_config.json")
dictConfig(config)

}
