% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib.R
\name{lib}
\alias{lib}
\alias{install_scriptR}
\alias{lib}
\title{Load Library Functions}
\usage{
install_scriptR()

lib(...)
}
\description{
Functions to load libraries
}
\details{
install_scriptR() installs the scriptR package \cr

lib() imports libraries depending on args: \cr
 data: (data wrangling) tidyverse, data.table, etc \cr
 develop: (development) argparse, optparse, logging \cr
 viz: (data visualization) ggplot2 and extensions \cr
}
