% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{logging}
\alias{logging}
\alias{logging.log}
\alias{logging.debug}
\alias{logging.info}
\alias{logging.warning}
\alias{logging.error}
\alias{logging.critical}
\alias{logging.exception}
\title{Python Style Logging Functions}
\usage{
logging.log(msg, level = "INFO")

logging.debug(x)

logging.info(x)

logging.warning(x)

logging.error(x)

logging.critical(x)

logging.exception(msg, e)
}
\description{
* logging.log() \cr
* logging.debug() \cr
* logging.info() \cr
* logging.warning() \cr
* logging.error() \cr
* logging.critical() \cr
* logging.exception()
}
