% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logmisc.R
\name{logmisc}
\alias{logmisc}
\title{Logs Misc Data Structures}
\usage{
logmisc(x)
}
\arguments{
\item{x}{any object to be logged}
}
\description{
Allows for logging of data structures such as dataframes, etc. Logs at the level INFO.
}
\examples{
logmisc(iris)
logmisc(summary(iris))

}
