% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_cmd_args.R
\name{print_cmd_args}
\alias{print_cmd_args}
\title{Prints Command Line Args}
\usage{
print_cmd_args(args)
}
\arguments{
\item{args}{list of command line args, generated from argparse::parse_args}
}
\description{
Logs out all command line arguments at the INFO level.
}
\examples{
library(argparse)

start_logging()

description <- "TEST"
parser <- ArgumentParser(description = description)
parser$add_argument("--test_opt", help = "This is a test")
parser$add_argument("--log", default = NULL, help = "Name of log config file [$(default)]")
args <- parser$parse_args()

print_cmd_args(args)

}
