% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_args.R
\name{process_args}
\alias{process_args}
\title{Process Command Line Args and Initiate Logging}
\usage{
process_args(parser, description, log_level = "DEBUG", ...)
}
\arguments{
\item{parser}{an argparse object}

\item{description}{description of script}

\item{...}{additional options passed to parse_args() call}
}
\value{
a list parsed options
}
\description{
Processes command line args, logs script description & arguments passed in. Returns list of option values.
}
\examples{

description <- "TEST"
parser <- ArgumentParser(description = description)
parser$add_argument("--test_opt", help = "This is a test")
parser$add_argument("--log", default = NULL, help = "Prefix of the log file; call be NULL or a default name [default]")

args <- process_options(parser = parser, description = description, log_level = "DEBUG", ...)

}
