% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.in.R
\name{read.in}
\alias{read.in}
\title{Opens a File}
\usage{
read.in(file, f = data.table::fread, ...)
}
\arguments{
\item{file}{character, file path}

\item{f}{function, reads in the file, defaults to data.table::fread}

\item{...}{additional arguments passed to f}
}
\description{
Wrapper around a read in data frame in a file (like data.table::fread or
read.table). Checks whether a file exists first and opens it with specified
function. Logs action at the INFO level.
}
\examples{
\dontrun{
read.in("test.csv")
read.in("dat.csv", read.table, header = TRUE)
}

}
