% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.in.R
\name{read.in}
\alias{read.in}
\title{Opens a File}
\usage{
read.in(filename, f = data.table::fread, ...)
}
\arguments{
\item{filename}{character, file path}

\item{f}{function, reads in the file, defaults to data.table::fread}

\item{...}{additional arguments passed to f}
}
\description{
Wrapper around a read in file (like data.table::fread or read.table). Checks
whether a file exists first and opens it with specified function
}
\examples{
read.in("test.csv")
read.in("dat.csv", read.table, header = TRUE)

}
