% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporters.R
\name{report}
\alias{report}
\alias{report_function_name}
\alias{report_args}
\alias{report_dim}
\alias{report_na}
\title{Reporter Functions}
\usage{
report_function_name(func)

report_args(func)

report_dim(func)

report_na(func)
}
\arguments{
\item{func}{function}
}
\value{
function
}
\description{
Wrappers to log various things\cr
* report_function_name() logs function name at the INFO level\cr
* report_args() logs function arguments at the DEBUG level\cr
* report_dim() logs data frame dimensions before and after function at the INFO level \cr
* report_na() logs number of na per column of the data frame at the DEBUG level
}
\details{
report_dim()'s function must take a data.frame as the first argument and return a data.frame
}
\examples{
start_logging()

this_is_a_function <- report_function_name(function(x) "Hi, I'm a function")
this_is_a_function()

my_function <- report_args(function(x, y, z) "Hi, I'm a function")
my_function(x = 1:5, y = list(a = letters[1:3], b = 'b'), z = head(iris))

head2 <- report_dim(head)
head2(mtcars)

}
