% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scriptr-package.R
\docType{package}
\name{scriptr-package}
\alias{scriptr}
\alias{scriptr-package}
\title{scriptr: Framework for storing code chunks and generating R scripts from them}
\description{
The package provides a framework for storing code chunks and generating R scripts from them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bundfussr.github.io/scriptr/}
  \item \url{https://github.com/bundfussr/scriptr}
}

}
\author{
\strong{Maintainer}: Stefan Bundfuss \email{stefan.bundfuss@roche.com} (\href{https://orcid.org/0009-0005-0027-1198}{ORCID})

}
\keyword{internal}
