% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_logging.R
\name{start_logging}
\alias{start_logging}
\title{Initiates a Logging Session}
\usage{
start_logging(file_name = NULL, log_level = "DEBUG")
}
\arguments{
\item{file_name}{name of the log file, if desired}

\item{log_level}{logging level, see logging::loglevels for available options}
}
\description{
Logs messages to command line, with option to create a log file.
}
\examples{
start_logging("log_test")

}
