% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_logging.R
\name{start_logging}
\alias{start_logging}
\title{Initiates a Logging Session}
\usage{
start_logging(file_name = NULL, config = NULL)
}
\arguments{
\item{file_name}{name of the log file, if desired}

\item{config}{name of the config file, json format}
}
\description{
Logs messages to command line, with option to create a log file or load configurations from a json file.
}
\details{
If config file not specified, will initiate a basic session (with logfile if specified).
}
\examples{
start_logging("log_test")

}
