% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_logging.R
\name{start_logging}
\alias{start_logging}
\title{Initiates a Logging Session}
\usage{
start_logging(config = NULL)
}
\arguments{
\item{config}{name of the config file, json format}
}
\description{
Logs messages to command line, load configurations from a json file.
}
\details{
If config file not specified, will initiate a basic session from the default config file.
}
\examples{
start_logging("log_config.json")

}
